Rename test folders and csproj removing spaces in name
