+++ .vscode/tasks.json
@@ -13,16 +13,16 @@
     // the command is a shell script
     "isShellCommand": true,
     // show the output window only if unrecognized errors occur.
-    "showOutput": "silent",
+    "showOutput": "always",
     // we run the custom script "compile" as defined in package.json
     "args": [
         "run",
         "compile",
         "--loglevel",
-        "silent"
+        ""
     ],
     // The tsc compiler is started in watching mode
-    "isBackground": true,
+    "isWatching": true,
     // use the standard tsc in watch mode problem matcher to find compile problems in the output.
     "problemMatcher": "$tsc-watch"
 }
