+++ test/test-dotnet-project/.vscode/settings.json
@@ -1,3 +1,3 @@
 {
-    "dotnet-test-explorer.testProjectPath": "unit-tests/xunit-tests/"
+    "dotnet-test-explorer.testProjectPath": "unittests/xunittests/"
 }
