+++ test/test-dotnet-project/test-dotnet-project.sln
@@ -7,11 +7,11 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{04BABC2C-CE7
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "testconsoleapp", "src\testconsoleapp\testconsoleapp.csproj", "{74CE9D6F-6F5F-4BA8-A8C8-9D69DB5AB19D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "testconsoleapp.tests", "unit-tests\xunit-tests\testconsoleapp-xunit.tests.csproj", "{95798EC8-0343-436B-9005-2E7814900000}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "testconsoleapp.tests", "unittests\xunittests\testconsoleappxunit.tests.csproj", "{95798EC8-0343-436B-9005-2E7814900000}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "unit-tests", "unit-tests", "{438F6A73-A30F-426F-873B-383D160D281D}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "unittests", "unittests", "{438F6A73-A30F-426F-873B-383D160D281D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "testconsoleapp-mstest.tests", "unit-tests\mstest-tests\testconsoleapp-mstest.tests.csproj", "{6A2370FE-2DBF-4BE2-81EE-D4746B723F88}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "testconsoleappmstest.tests", "unittests\mstesttests\testconsoleappmstest.tests.csproj", "{6A2370FE-2DBF-4BE2-81EE-D4746B723F88}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
