+++ test/test-dotnet-project/unittests/xunittests/testconsoleappxunit.tests.csproj
@@ -13,6 +13,7 @@
   </ItemGroup>
 
   <ItemGroup>
+
     <ProjectReference Include="..\..\src\testconsoleapp\testconsoleapp.csproj" />
   </ItemGroup>
